pyproject.toml
[project]
name = "labdados"
version = "0.1.0"
description = "SDK Python para os serviços do escritório de apoio do LabDados / FGV Direito SP — OCR, transcrição, estruturação com LLMs e análise de viabilidade. Roda em modo nuvem (com API key) ou local."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "LabDados / FGV Direito SP", email = "labdados.direitosp@fgv.br" },
]
requires-python = ">=3.10"
keywords = [
    "labdados",
    "fgv",
    "direito",
    "ocr",
    "transcrição",
    "whisper",
    "llm",
    "estruturação",
    "viabilidade",
    "tribunais",
    "datajud",
    "jurisprudência",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    # Único requisito do modo nuvem — usado para chamar a API REST e baixar
    # arquivos do Azure Blob via SAS URL.
    "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
# OCR local: PyMuPDF abre o PDF e renderiza páginas; pytesseract chama o
# binário do Tesseract (instalado pelo OS). Suficiente pra processar de
# forma decente sem GPU.
ocr = [
    "pymupdf>=1.24",
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
]
# Transcrição local: faster-whisper roda em CPU ou GPU CUDA (se disponível).
# WhisperX (com diarização) tem deps pesadas e fica fora do extra "leve" —
# documentado mas não é instalado automaticamente.
transcricao = [
    "faster-whisper>=1.0",
]
# Estruturação local: cliente openai (compatível com Azure OpenAI, Ollama
# e qualquer servidor OpenAI-compat). O usuário traz a própria chave.
estruturacao = [
    "openai>=1.40",
]
# Análise de viabilidade local: juscraper consulta tribunais (Datajud +
# bancos de jurisprudência); jinja2 + (Quarto via OS) renderiza relatório.
viabilidade = [
    "juscraper>=0.4",
    "jinja2>=3.1",
    "httpx>=0.27,<1.0",
]
all = [
    "labdados[ocr,transcricao,estruturacao,viabilidade]",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "respx>=0.21",
]
docs = [
    "quartodoc>=0.7",
    "griffe>=0.45",
]

[project.urls]
Homepage = "https://escritorio.labdados.fgv.br"
Documentation = "https://labdados.github.io/labdados-sdk"
Repository = "https://github.com/labdados/labdados-sdk"
Issues = "https://github.com/labdados/labdados-sdk/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/labdados"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
